finished refactoring node.py, glossary.py
